Memory and Usability improvements
using inheritance with virtual methods consumes "quite a lot" of
memory: The vtable to store the method of a class and the class pointer
of each object need to be stored somewhere. Of course, this is just a
few bytes, but in tight memory sitations as on an arduino, every byte
counts.

To allow saving space, the actual sensor handling was moved to a new,
stripped down class that's just doing the bare minimum (Sensor_TSL2591).

To keep the API compatible (for cases where memory doesn't matter), the
old Adafruit_TSL2591 class is turned into a wrapper. Memory consumption
is the same using the Wrapper ... compared to the former implementation.

This wrapper is also hiding some behavior changes of the new
Sensor_TSL2591 Class:
- most methods now check for detected errors and report them.
- return values now allow reliable detaction and handling of errors.
- the sensor isn't disabled automatically
- getFullLuminosity won't wait for the integration time.

The last two changes allow more advanced power management when using the
stripped down class directly.

Furthermore, getFullLuminosity now reads the whole data in one go to
avoid the individual channels to become corrupted while reading ... as
suggested by the datasheet.